refactor text file handling
